% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_interventions.R
\name{assess_interventions}
\alias{assess_interventions}
\title{Assess different intervention strategies}
\usage{
assess_interventions(
  koala_model,
  years,
  frequency = c(1, 2, 3, 4),
  proportion = seq(0, 1, by = 0.05),
  cl = NULL,
  ...
)
}
\arguments{
\item{koala_model}{a valid \code{\link{KoalasV2}} model}

\item{years}{the number of years to update the model (passed to the run method of the \code{\link{KoalasV2}} model)}

\item{frequency}{a vector of annual sampling frequencies to test}

\item{proportion}{a vector of sampling proportions to test}

\item{cl}{passed to \code{\link[pbapply]{pblapply}} - DO NOT CHANGE THIS ON WINDOWS as it won't properly without shared memory forking}

\item{...}{additional arguments (cull proportions) passed to the active_intervention method of the \code{\link{KoalasV2}} model}
}
\value{
a data frame of final prevalence and population size values at the end of each simulation for each combination of frequency and proportion
}
\description{
Assess different intervention strategies
}
