% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoalasR6.R
\name{KoalasR6}
\alias{KoalasR6}
\title{R6 representation of the Koala model}
\description{
This is a within-group SVIDR(D) model class that can either be run on its own
or embedded in a between-group model.
Note: D is not a real state (and can be negative), it is just used to ensure
that the books are balanced including birth/death
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{S}}{number of susceptible animals}

\item{\code{I}}{total number of infected animals}

\item{\code{D}}{number of diseased animals}

\item{\code{R}}{number of recovered animals}

\item{\code{V}}{number of vaccinated animals}

\item{\code{Pt}}{cumulative number of animals removed as true positives on active sampling}

\item{\code{Pf}}{cumulative number of animals removed as false positives on active sampling}

\item{\code{Pc}}{cumulative number of animals removed as true positives on passive sampling}

\item{\code{mortality_natural}}{mortality rate from natural causes}

\item{\code{mortality_disease}}{mortality rate from disease}

\item{\code{carrying_capacity}}{number of animals supported by the population (can be Inf or NA)}

\item{\code{birth_rate}}{birth rate}

\item{\code{relative_fecundity}}{fecundity of diseased animals relative to other states (S/I/R/V)}

\item{\code{beta}}{the transmission rate parameter per unit time (must be positive)}

\item{\code{sigma}}{the progression rate from I to D or S}

\item{\code{acute_recovery_prob}}{the probability that a progression from I will go to S rather than D}

\item{\code{recovery}}{the recovery rate parameter per unit time (must be positive)}

\item{\code{waning_natural}}{the immune waning rate from R (following natural infection)}

\item{\code{waning_vaccine}}{the immune waning rate from V (following vaccination)}

\item{\code{time}}{the current time point of the model (read-only)}

\item{\code{N}}{the total number of animals alive (read-only)}

\item{\code{state}}{a data frame representing the current state of the model (read-only)}

\item{\code{trans_external}}{the external transmission parameter}

\item{\code{transmission_type}}{either frequency or density}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KoalasR6-new}{\code{KoalasR6$new()}}
\item \href{#method-KoalasR6-update}{\code{KoalasR6$update()}}
\item \href{#method-KoalasR6-run}{\code{KoalasR6$run()}}
\item \href{#method-KoalasR6-vaccinate}{\code{KoalasR6$vaccinate()}}
\item \href{#method-KoalasR6-active_test}{\code{KoalasR6$active_test()}}
\item \href{#method-KoalasR6-passive_test}{\code{KoalasR6$passive_test()}}
\item \href{#method-KoalasR6-print}{\code{KoalasR6$print()}}
\item \href{#method-KoalasR6-clone}{\code{KoalasR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-new}{}}}
\subsection{Method \code{new()}}{
Create a new within-group model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$new(
  update_type = c("deterministic", "stochastic"),
  group_name = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{update_type}}{one of deterministic or stochastic}

\item{\code{group_name}}{an optional name for the group (will be included in the output state, if provided)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new within-group model object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-update"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-update}{}}}
\subsection{Method \code{update()}}{
Update the state of the group for a single time point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$update(d_time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d_time}}{the desired time step (delta time)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-run"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-run}{}}}
\subsection{Method \code{run()}}{
Update the state of the group for several time points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$run(add_time, d_time, thin = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_time}}{the additional time to add to the current time of the model}

\item{\code{d_time}}{the desired time step (delta time)}

\item{\code{thin}}{thinning parameter (currently ignored)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of the model state at each (new) time point
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-vaccinate"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-vaccinate}{}}}
\subsection{Method \code{vaccinate()}}{
Implement a (one-time) vaccination effort at the current time point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$vaccinate(proportion, number, efficacy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proportion}}{the proportion of animals to vaccinate (ignored if number is supplied)}

\item{\code{number}}{the (maximum) number of animals to vaccinate}

\item{\code{efficacy}}{the efficacy of the vaccine i.e. probability of the animal moving to V (can be a vector with names S/I/D/R/V, otherwise the supplied number is re-used except for 0 is used for R)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the total number of animals vaccinated (including any R)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-active_test"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-active_test}{}}}
\subsection{Method \code{active_test()}}{
Implement a (one-time) active sampling/capture/testing of all animals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$active_test(proportion, number, sensitivity, specificity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proportion}}{the proportion of animals to test and remove (ignored if number is supplied)}

\item{\code{number}}{the (maximum) number of animals to test and remove}

\item{\code{sensitivity}}{the sensitivity of the test}

\item{\code{specificity}}{the specificity of the test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a 2x2 matrix with number of Tp/Fn/Tn/Fp tests
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-passive_test"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-passive_test}{}}}
\subsection{Method \code{passive_test()}}{
Implement a (one-time) passive sampling/capture/testing of diseased animals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$passive_test(proportion, number, sensitivity, specificity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proportion}}{the proportion of diseased animals to test and remove (ignored if number is supplied)}

\item{\code{number}}{the (maximum) number of animals to test and remove}

\item{\code{sensitivity}}{the sensitivity of the test}

\item{\code{specificity}}{the specificity of the test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a 2x2 matrix with number of Tp/Fn/Tn/Fp tests (although Tn and Fp are zero)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-print"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-print}{}}}
\subsection{Method \code{print()}}{
Print method giving an overview of the current state and parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
