% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoalasV2.R
\name{KoalasV2}
\alias{KoalasV2}
\title{R6 representation of the Koala model}
\description{
This is a within-group SVIDR(D) model class that can either be run on its own
or embedded in a between-group model.
Note: D is not a real state (and can be negative), it is just used to ensure
that the books are balanced including birth/death
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{S}}{number of susceptible animals}

\item{\code{I}}{total number of infected animals}

\item{\code{D}}{number of diseased animals}

\item{\code{R}}{number of recovered animals}

\item{\code{V}}{number of vaccinated animals}

\item{\code{Ctp}}{cumulative number of animals removed as true positives on active/passive sampling}

\item{\code{Cfp}}{cumulative number of animals removed as false positives on active/passive sampling}

\item{\code{mortality_natural}}{mortality rate from natural causes}

\item{\code{mortality_disease}}{mortality rate from disease}

\item{\code{carrying_capacity}}{number of animals supported by the population (can be Inf or NA)}

\item{\code{birth_rate}}{birth rate}

\item{\code{relative_fecundity}}{fecundity of diseased animals relative to other states (S/I/R/V)}

\item{\code{beta}}{the transmission rate parameter per unit time (must be positive)}

\item{\code{sigma}}{the progression rate from I to D or S}

\item{\code{acute_recovery_prob}}{the probability that a progression from I will go to S rather than D}

\item{\code{recovery}}{the recovery rate parameter per unit time (must be positive)}

\item{\code{waning_natural}}{the immune waning rate from R (following natural infection)}

\item{\code{waning_vaccine}}{the immune waning rate from V (following vaccination)}

\item{\code{time}}{the current time point of the model (read-only)}

\item{\code{N}}{the total number of animals alive (read-only)}

\item{\code{state}}{a data frame representing the current state of the model (read-only)}

\item{\code{trans_external}}{the external transmission parameter}

\item{\code{transmission_type}}{either frequency or density}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KoalasV2-new}{\code{KoalasV2$new()}}
\item \href{#method-KoalasV2-update}{\code{KoalasV2$update()}}
\item \href{#method-KoalasV2-run}{\code{KoalasV2$run()}}
\item \href{#method-KoalasV2-vaccinate}{\code{KoalasV2$vaccinate()}}
\item \href{#method-KoalasV2-active_test}{\code{KoalasV2$active_test()}}
\item \href{#method-KoalasV2-passive_test}{\code{KoalasV2$passive_test()}}
\item \href{#method-KoalasV2-print}{\code{KoalasV2$print()}}
\item \href{#method-KoalasV2-clone}{\code{KoalasV2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-new"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-new}{}}}
\subsection{Method \code{new()}}{
Create a new within-group model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$new(num_V = 3L, num_I = 3L, num_N = 3L, num_R = 3L, num_A = 3L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_V}}{number of sub-components for V (and Vf)}

\item{\code{num_I}}{number of sub-components for I (and If)}

\item{\code{num_N}}{number of sub-components for N (and Nf)}

\item{\code{num_R}}{number of sub-components for R (and Rf)}

\item{\code{num_A}}{number of sub-components for A (and If)}

\item{\code{group_name}}{an optional name for the group (will be included in the output state, if provided)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new within-group model object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-update"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-update}{}}}
\subsection{Method \code{update()}}{
Update the state of the group for a single time point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$update(d_time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d_time}}{the desired time step (delta time)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-run"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-run}{}}}
\subsection{Method \code{run()}}{
Update the state of the group for several time points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$run(add_time, d_time, thin = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_time}}{the additional time to add to the current time of the model}

\item{\code{d_time}}{the desired time step (delta time)}

\item{\code{thin}}{thinning parameter (currently ignored)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of the model state at each (new) time point
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-vaccinate"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-vaccinate}{}}}
\subsection{Method \code{vaccinate()}}{
Implement a (one-time) vaccination effort at the current time point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$vaccinate(number, proportion, efficacy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{the (maximum) number of animals to vaccinate (ignored if proportion is supplied)}

\item{\code{proportion}}{the proportion of animals to vaccinate}

\item{\code{efficacy}}{the efficacy of the vaccine i.e. probability of the animal moving to V (from S, I or D - R and V do not move)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-active_test"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-active_test}{}}}
\subsection{Method \code{active_test()}}{
Implement a (one-time) active sampling/capture/testing of all animals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$active_test(number, proportion, sensitivity, specificity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{the (maximum) number of animals to test and remove (ignored if proportion is supplied)}

\item{\code{proportion}}{the proportion of animals to test and remove}

\item{\code{sensitivity}}{the sensitivity of the test}

\item{\code{specificity}}{the specificity of the test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-passive_test"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-passive_test}{}}}
\subsection{Method \code{passive_test()}}{
Implement a (one-time) passive sampling/capture/testing of diseased animals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$passive_test(
  number,
  proportion,
  sensitivity,
  specificity,
  prevalence_other
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{the (maximum) number of animals to test and remove (ignored if proportion is supplied)}

\item{\code{proportion}}{the proportion of animals to test and remove}

\item{\code{sensitivity}}{the sensitivity of the test}

\item{\code{specificity}}{the specificity of the test}

\item{\code{prevalence_other}}{the prevalence of other conditions that resemble clinical disease}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-print"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-print}{}}}
\subsection{Method \code{print()}}{
Print method giving an overview of the current state and parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasV2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasV2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasV2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
