% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KoalasR6.R
\name{KoalasR6}
\alias{KoalasR6}
\title{R6 representation of the Koala model}
\description{
This is a within-group SVIDR(D) model class that can either be run on its own
or embedded in a between-group model.
Note: D is not a real state (and can be negative), it is just used to ensure
that the books are balanced including birth/death
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{omega}}{the progression rate from I to D or S}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{S}}{number of susceptible animals}

\item{\code{I}}{total number of infected animals}

\item{\code{D}}{number of diseased animals}

\item{\code{R}}{number of recovered animals}

\item{\code{mortality_natural}}{mortality rate from natural causes}

\item{\code{mortality_disease}}{mortality rate from disease}

\item{\code{carrying_capacity}}{number of animals supported by the population (can be Inf or NA)}

\item{\code{birth_rate}}{birth rate}

\item{\code{relative_fecundity}}{fecundity of diseased animals relative to other states (S/I/R/V)}

\item{\code{beta}}{the transmission rate parameter per unit time (must be positive)}

\item{\code{omega}}{the progression rate from I to D or S}

\item{\code{acute_recovery_prob}}{the probability that a progression from I will go to S rather than D}

\item{\code{recovery}}{the recovery rate parameter per unit time (must be positive)}

\item{\code{waning_natural}}{the immune waning rate from R (following natural infection)}

\item{\code{waning_vaccine}}{the immune waning rate from V (following vaccination)}

\item{\code{time}}{the current time point of the model (read-only)}

\item{\code{N}}{the total number of animals alive (read-only)}

\item{\code{state}}{a data frame representing the current state of the model (read-only)}

\item{\code{trans_external}}{the external transmission parameter}

\item{\code{transmission_type}}{either frequency or density}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KoalasR6-new}{\code{KoalasR6$new()}}
\item \href{#method-KoalasR6-update}{\code{KoalasR6$update()}}
\item \href{#method-KoalasR6-run}{\code{KoalasR6$run()}}
\item \href{#method-KoalasR6-print}{\code{KoalasR6$print()}}
\item \href{#method-KoalasR6-clone}{\code{KoalasR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-new}{}}}
\subsection{Method \code{new()}}{
Create a new within-group model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$new(
  update_type = c("deterministic", "stochastic"),
  group_name = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{update_type}}{one of deterministic or stochastic}

\item{\code{group_name}}{an optional name for the group (will be included in the output state, if provided)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new within-group model object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-update"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-update}{}}}
\subsection{Method \code{update()}}{
Update the state of the group for a single time point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$update(d_time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d_time}}{the desired time step (delta time)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-run"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-run}{}}}
\subsection{Method \code{run()}}{
Update the state of the group for several time points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$run(add_time, d_time, thin = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_time}}{the additional time to add to the current time of the model}

\item{\code{d_time}}{the desired time step (delta time)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame of the model state at each (new) time point
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-print"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-print}{}}}
\subsection{Method \code{print()}}{
Print method giving an overview of the current state and parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KoalasR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KoalasR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoalasR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
